Store the code-snippet ID as four bytes.
If we are going to create a new snippet ID every time a runtime effect
is created, we will need the code-snippet ID to support values larger
than a byte.

Bug: skia:13405
Reviewed-on: https://skia-review.googlesource.com/c/skia/+/551761
Auto-Submit: John Stiles <johnstiles@google.com>
Commit-Queue: John Stiles <johnstiles@google.com>

+++ src/core/SkBuiltInCodeSnippetID.h
@@ -11,7 +11,7 @@
 #include "include/core/SkTypes.h"
 
 // TODO: this needs to be expanded into a more flexible dictionary (esp. for user-supplied SkSL)
-enum class SkBuiltInCodeSnippetID : uint8_t {
+enum class SkBuiltInCodeSnippetID : int32_t {
     // This isn't just a signal for a failure during paintparams key creation. It also actually
     // implements the default behavior for an erroneous draw. Currently it just draws solid
     // magenta.
